Update threshold to mapped value

+++ LightSensor/src/LightSensor.cpp
@@ -14,13 +14,13 @@ namespace
 LightSensor::LightSensor(uint8_t analogPin, Sensitivity sensitivity)
     : IObservable(1)
     , m_pin(analogPin)
-    , m_threshold(static_cast<uint16_t>(sensitivity))
+    , m_threshold(static_cast<uint8_t>(sensitivity))
 {
 }
 
 void LightSensor::HandleEvents()
 {
-    const auto Difference = [](uint16_t valueOne, uint16_t valueTwo){ return (valueOne >= valueTwo) ? (valueOne - valueTwo) : (valueTwo - valueOne); };
+    const auto Difference = [](uint8_t valueOne, uint8_t valueTwo){ return (valueOne >= valueTwo) ? (valueOne - valueTwo) : (valueTwo - valueOne); };
 
     if (const auto currentValue = Value(); Difference(currentValue, m_value) >= m_threshold)
     {
